data-pipeline-1/requirements.txt
pandas
pytest
python-dotenv
pymssql
requests
